Integrate database in the backend and separate error handling.
Integrate our database into the mix and refactor the code accordingly for
all routes. Ensure that whenever new contact is posted, we perform a check
to see if a contact with the same name exists or not. If it does then only
update it. Also, extract error handling into a separate function which we
utilize through the catch blocks in each promise chain.

Separate out our schema and model configuration and connection to the
database into its own module and import it into the index.js to separate
concerns.