First draft of bloglist project
Take template code for index.js from the course and created
separated modules for more scalable code. Separate utils such as
middleware and config(for initiation of all environment variables)
as well as the blog model and blogs controller in their respective
folders.

Integrated all the parts into the app.js file and exported it to
index.js. Note: have not implemented logger util yet and will do it
right after this commit.